+++ src/index.test.js
@@ -41,13 +41,13 @@ describe('jss-expand', () => {
     })
   })
 
-  describe('comma-separated values as arrays', () => {
+  describe('comma-separated values as arrays (using double arrays)', () => {
     let sheet
 
     beforeEach(() => {
       sheet = jss.createStyleSheet({
         a: {
-          transition: ['opacity 1s linear', 'transform 300ms ease']
+          transition: [['opacity', 1, 'linear'], ['transform', 300, 'ease']]
         }
       })
     })
@@ -59,7 +59,7 @@ describe('jss-expand', () => {
     it('should generate correct CSS', () => {
       expect(sheet.toString()).to.be(
         '.a-id {\n' +
-        '  transition: opacity 1s linear,transform 300ms ease;\n' +
+        '  transition: opacity 1 linear, transform 300 ease;\n' +
         '}'
       )
     })
@@ -249,6 +249,63 @@ describe('jss-expand', () => {
       expect(sheet.toString()).to.be(
         '.a-id {\n' +
         '  transition: opacity 200ms linear 300ms;\n' +
+        '}'
+      )
+    })
+  })
+
+  describe('non-standart properties support', () => {
+    let sheet
+
+    beforeEach(() => {
+      sheet = jss.createStyleSheet({
+        a: {
+          border: {
+            width: '2px',
+            style: 'solid',
+            color: 'black',
+            radius: ['5px', '10px']
+          }
+        }
+      })
+    })
+
+    it('should add rules', () => {
+      expect(sheet.getRule('a')).to.not.be(undefined)
+    })
+
+    it('should generate correct CSS', () => {
+      expect(sheet.toString()).to.be(
+        '.a-id {\n' +
+        '  border: 2px solid black;\n' +
+        '  border-radius: 5px 10px;\n' +
+        '}'
+      )
+    })
+  })
+
+  describe('non-standart properties should not overwrite standart properties notation', () => {
+    let sheet
+
+    beforeEach(() => {
+      sheet = jss.createStyleSheet({
+        a: {
+          border: {
+            radius: ['5px', '10px']
+          },
+          'border-radius': '10px'
+        }
+      })
+    })
+
+    it('should add rules', () => {
+      expect(sheet.getRule('a')).to.not.be(undefined)
+    })
+
+    it('should generate correct CSS', () => {
+      expect(sheet.toString()).to.be(
+        '.a-id {\n' +
+        '  border-radius: 10px;\n' +
         '}'
       )
     })